# Optimize persistent connection registry for 8-10% performance gains