--- a/lib/ezclient/client.rb
+++ b/lib/ezclient/client.rb
@@ -32,13 +32,13 @@ class EzClient::Client
 
     if keep_alive_timeout
       client = persistent_client_registry.for(
-        url, ssl_context: ssl_context, timeout: keep_alive_timeout
+        url, ssl_context:, timeout: keep_alive_timeout
       )
     else
       client = HTTP::Client.new
     end
 
-    EzClient::Request.new(verb, url, client: client, **options).tap do |request|
+    EzClient::Request.new(verb, url, client:, **options).tap do |request|
       request.api_auth!(*api_auth) if api_auth
     end
   end