--- a/lib/ezclient/request.rb
+++ b/lib/ezclient/request.rb
@@ -115,11 +115,11 @@ class EzClient::Request
     self.elapsed_seconds = EzClient.get_time - perform_started_at
   end
 
-  def with_retry(&block)
+  def with_retry(&)
     retries = 0
 
     begin
-      retry_on_connection_error(&block)
+      retry_on_connection_error(&)
     rescue *retried_exceptions => error
       if retries < max_retries.to_i
         retries += 1
@@ -209,7 +209,7 @@ class EzClient::Request
       case options[:basic_auth]
       when Array
         user, password = options[:basic_auth]
-        { user: user, pass: password }
+        { user:, pass: password }
       when Hash
         options[:basic_auth]
       end