--- a/lib/ezclient/version.rb
+++ b/lib/ezclient/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module EzClient
-  VERSION = "1.7.2"
+  VERSION = "1.7.3"
 end