--- a/spec/ezclient_spec.rb
+++ b/spec/ezclient_spec.rb
@@ -52,7 +52,7 @@ RSpec.describe EzClient do
     end
 
     context "when headers request option is provided" do
-      let(:request_options) { { headers: headers } }
+      let(:request_options) { { headers: } }
       let(:headers) { { some_header: 1 } }
 
       it "makes request with proper headers" do
@@ -78,7 +78,7 @@ RSpec.describe EzClient do
       end
 
       context "when cookies request option is provided" do
-        let(:request_options) { { headers: headers, cookies: cookies } }
+        let(:request_options) { { headers:, cookies: } }
         let(:cookies) { { a: 1 } }
 
         it "makes request with proper headers" do
@@ -138,7 +138,7 @@ RSpec.describe EzClient do
     end
 
     context "when params request option is provided" do
-      let(:request_options) { { params: params } }
+      let(:request_options) { { params: } }
       let(:params) { { a: 1 } }
 
       it "makes proper request" do
@@ -185,7 +185,7 @@ RSpec.describe EzClient do
     end
 
     context "when on_complete callback is provided" do
-      let(:client_options) { { on_complete: on_complete } }
+      let(:client_options) { { on_complete: } }
       let(:calls) { [] }
 
       let(:on_complete) do
@@ -253,7 +253,7 @@ RSpec.describe EzClient do
     end
 
     context "when on_error callback is provided" do
-      let(:client_options) { { on_error: on_error } }
+      let(:client_options) { { on_error: } }
       let(:calls) { [] }
 
       let(:on_error) do
@@ -273,7 +273,7 @@ RSpec.describe EzClient do
     end
 
     context "when error_wrapper callback is provided" do
-      let(:client_options) { { error_wrapper: error_wrapper } }
+      let(:client_options) { { error_wrapper: } }
       let(:calls) { [] }
 
       let(:error_wrapper) do
@@ -302,7 +302,7 @@ RSpec.describe EzClient do
     end
 
     context "when on_retry callback is provided" do
-      let(:client_options) { { on_retry: on_retry } }
+      let(:client_options) { { on_retry: } }
       let(:request_options) { { metadata: :smth } }
       let(:calls) { [] }
 
@@ -511,7 +511,7 @@ RSpec.describe EzClient do
       end
 
       context "when on_retry callback is provided" do
-        let(:client_options) { { on_retry: on_retry } }
+        let(:client_options) { { on_retry: } }
         let(:request_options) { { retry_exceptions: SomeError, metadata: :smth } }
         let(:calls) { [] }
 
@@ -570,7 +570,7 @@ RSpec.describe EzClient do
       ssl_context.cert = cert
 
       client.perform!(:get, "https://example1.com")
-      client.perform!(:get, "https://example2.com", ssl_context: ssl_context)
+      client.perform!(:get, "https://example2.com", ssl_context:)
 
       registry.send(:get_origin, "https://example3.com/path")