--- a/spec/persistent_connections_spec.rb
+++ b/spec/persistent_connections_spec.rb
@@ -199,6 +199,5 @@ RSpec.describe "Persistent Connections" do
         end
       end
     end
-
   end
 end