## Overview

1. Certificate Hash Caching (8.3-8.7% improvement)

  - Implements SHA256 hash caching with `compare_by_identity` for SSL certificates
  - Avoids repeated expensive cryptographic operations

 2. URL Origin Caching (9.9-10.6% improvement)

  - Caches parsed URL origins to avoid repeated URI parsing
  - Some speedup for applications making multiple requests to the same domains

  3. Registry Structure Optimization

  - Replaces nested hash structure `registry[origin][ssl_bucket]` with flat compound keys
  - Simplifies lookups and reduces memory overhead

  4. Cleanup Interval Optimization (8.16% improvement)

  - Changes from immediate cleanup on every request to 60-second intervals
  - Reduces overhead while maintaining reasonable memory usage
  - Handle cases with time outed conns before triggered cleanup

 5. Update minimum supported Ruby version to: `3.1`: https://www.ruby-lang.org/en/downloads/branches

## Benchmarks 

Look at `bin/benchmark` executable file.

```
============================================================
COMPARISON WITH MASTER
============================================================

Ssl certificate requests:
  requests with same SSL cert:
    Master:  88.84 i/s
    Current: 96.21 i/s
    Change:  8.3% 🚀 FASTER
  requests with different SSL certs:
    Master:  44.06 i/s
    Current: 47.89 i/s
    Change:  8.69% 🚀 FASTER

Multiple url requests:
  requests to same origin:
    Master:  90.9 i/s
    Current: 99.89 i/s
    Change:  9.9% 🚀 FASTER
  requests to different origins:
    Master:  57.27 i/s
    Current: 63.35 i/s
    Change:  10.62% 🚀 FASTER

Cleanup performance:
  frequent requests:
    Master:  133.72 i/s
    Current: 144.63 i/s
    Change:  8.16% 🚀 FASTER

Memory Usage:
  Total allocated bytes:
    Master:  4966980
    Current: 4761220
    Change:  -205760 bytes (-4.14%) ✅ BETTER
  Total retained bytes:
    Master:  18638
    Current: 18638
    Change:  0 bytes (0.0%) ❌ WORSE
```